Merges estimator_abfe.py and estimator.py
* Prefers estimator_abfe.py logic
